refactor(docs): separate family document rules
